+++ src/modules/network/generator.cpp
@@ -178,9 +178,9 @@ void generator::config(const model::generator_config& config)
 
     if (config.ratio) {
         t_config.synchronizer = &m_synchronizer;
-        m_synchronizer.ratio_reads.store(config.ratio.value().reads,
+        m_synchronizer.ratio_reads.store(static_cast<int32_t>(config.ratio.value().reads),
                                          std::memory_order_relaxed);
-        m_synchronizer.ratio_writes.store(config.ratio.value().writes,
+        m_synchronizer.ratio_writes.store(static_cast<int32_t>(config.ratio.value().writes),
                                           std::memory_order_relaxed);
     }
